apirest/requirements.txt
Django==1.11.13
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
